# Optimize aggregations in PromQL engine