I'm investigating some slow queries in Cortex (running the vanilla PromQL engine). In particular, I'm investigating some high cardinality queries (50k+ series) with many steps (5000) in the form of `sum by(...) (metric{...})`.

Profiling such queries I've noticed that (in my specific case) 60% of CPU is spent by `Labels.HashForLabels()`:

![Screenshot 2021-03-13 at 07 21 54](https://user-images.githubusercontent.com/1701904/111021513-3b381e80-83cd-11eb-853f-f1e755aa157a.png)

I believe there's a significant inefficiency regarding that. The `HashForLabels()` is called for every series **and** every step in order to generate the grouping key but the grouping key doesn't change on each step. In this PR I'm proposing to compute the grouping key for each series upfront and then pass it to each `aggregation()` call.

_In this PR I've intentionally not optimized `count_values()` because would be more complicated and, in my experience, unfrequently used compared to other aggregation operators._

## Benchmarks TL;DR

The focus in the benchmark should be on queries running aggregations, in particular the case "many steps". Queries are fast on 1 single step, but significantly slower on many steps. As you can see, the CPU reduction (and so execution time) is significant at the cost of a slight increase of allocated memory and allocations.

```
name                                                                                                       old time/op    new time/op    delta
RangeQuery/expr=sum(a_one),steps=1000-12                                                                      593µs ± 0%     584µs ± 0%   -1.53%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_ten),steps=1000-12                                                                     1.38ms ± 3%    1.32ms ± 0%   -4.75%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                 9.00ms ± 1%    8.66ms ± 0%   -3.82%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-12                                                         6.10ms ± 0%    5.75ms ± 0%   -5.70%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-12                                                         19.7ms ± 0%    15.0ms ± 1%  -24.00%  (p=0.016 n=4+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                      159ms ± 1%     119ms ± 2%  -24.79%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-12                                                        1.78ms ± 1%    1.49ms ± 1%  -16.43%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-12                                                        18.1ms ± 1%    14.1ms ± 0%  -22.35%  (p=0.016 n=4+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                     191ms ± 0%     159ms ± 2%  -17.13%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-12                                                              1.73ms ± 0%    1.43ms ± 1%  -17.21%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_ten),steps=1000-12                                                              17.1ms ± 0%    13.5ms ± 0%  -21.40%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                           186ms ± 1%     151ms ± 2%  -18.87%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_one),steps=1000-12                                                             5.60ms ± 0%    5.25ms ± 2%   -6.20%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_ten),steps=1000-12                                                             19.7ms ± 0%    14.3ms ± 3%  -27.21%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                          165ms ± 1%     118ms ± 1%  -28.43%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=1000-12                                                726µs ± 1%     688µs ± 2%   -5.23%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1000-12                                               2.35ms ± 0%    2.12ms ± 0%   -9.49%  (p=0.016 n=5+4)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                           17.9ms ± 1%    16.4ms ± 0%   -8.36%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1000-12            1.93ms ± 2%    1.85ms ± 0%   -4.11%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1000-12            5.15ms ± 0%    4.76ms ± 0%   -7.49%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12    36.4ms ± 2%    33.8ms ± 0%   -7.24%  (p=0.008 n=5+5)

name                                                                                                       old alloc/op   new alloc/op   delta
RangeQuery/expr=sum(a_one),steps=1000-12                                                                      442kB ± 0%     443kB ± 0%   +0.10%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_ten),steps=1000-12                                                                      472kB ± 0%     476kB ± 0%   +0.76%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                  764kB ± 0%     798kB ± 0%   +4.41%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-12                                                         2.96MB ± 0%    2.93MB ± 0%   -0.94%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-12                                                         3.65MB ± 0%    3.65MB ± 0%   +0.16%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                     6.97MB ± 0%    7.31MB ± 0%   +4.86%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-12                                                         571kB ± 0%     543kB ± 0%   -4.92%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-12                                                        3.41MB ± 0%    3.42MB ± 0%   +0.17%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                    33.4MB ± 0%    33.7MB ± 0%   +1.00%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-12                                                               539kB ± 0%     511kB ± 0%   -5.21%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_ten),steps=1000-12                                                              2.77MB ± 0%    2.78MB ± 0%   +0.22%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                          27.0MB ± 0%    27.3MB ± 0%   +1.26%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_one),steps=1000-12                                                             2.61MB ± 0%    2.58MB ± 0%   -1.07%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_ten),steps=1000-12                                                             2.94MB ± 0%    2.95MB ± 0%   +0.20%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                         6.26MB ± 0%    6.60MB ± 0%   +5.42%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=1000-12                                                476kB ± 0%     444kB ± 0%   -6.65%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1000-12                                                539kB ± 0%     510kB ± 0%   -5.44%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                            845kB ± 0%     838kB ± 0%   -0.81%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1000-12            1.11MB ± 0%    1.05MB ± 0%   -5.71%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1000-12            1.24MB ± 0%    1.18MB ± 0%   -4.72%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12    1.85MB ± 0%    1.84MB ± 0%   -0.68%  (p=0.008 n=5+5)

name                                                                                                       old allocs/op  new allocs/op  delta
RangeQuery/expr=sum(a_one),steps=1000-12                                                                      6.18k ± 0%     6.20k ± 0%   +0.34%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_ten),steps=1000-12                                                                      6.66k ± 0%     6.83k ± 0%   +2.61%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                  11.3k ± 0%     13.0k ± 0%  +14.64%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-12                                                          41.1k ± 0%     40.3k ± 0%   -1.97%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-12                                                          46.2k ± 0%     47.1k ± 0%   +1.89%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                      97.7k ± 0%    114.9k ± 0%  +17.60%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-12                                                         8.71k ± 0%     7.90k ± 0%   -9.29%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-12                                                         43.0k ± 0%     43.9k ± 0%   +2.03%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                      379k ± 0%      396k ± 0%   +4.54%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-12                                                               8.71k ± 0%     7.90k ± 0%   -9.29%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_ten),steps=1000-12                                                               43.0k ± 0%     43.9k ± 0%   +2.03%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                            379k ± 0%      396k ± 0%   +4.55%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_one),steps=1000-12                                                              41.1k ± 0%     40.3k ± 0%   -1.97%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_ten),steps=1000-12                                                              46.2k ± 0%     47.1k ± 0%   +1.89%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                          97.7k ± 0%    114.9k ± 0%  +17.61%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=1000-12                                                7.21k ± 0%     6.23k ± 0%  -13.63%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1000-12                                                8.69k ± 0%     7.82k ± 0%   -9.97%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                            13.4k ± 0%     13.6k ± 0%   +1.94%  (p=0.016 n=5+4)
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1000-12             18.4k ± 0%     16.4k ± 0%  -10.69%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1000-12             21.3k ± 0%     19.6k ± 0%   -8.11%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12     30.7k ± 0%     31.2k ± 0%   +1.69%  (p=0.016 n=4+5)
```

## Full benchmarks

```
name                                                                                                       old time/op    new time/op    delta
RangeQuery/expr=a_one,steps=1-12                                                                             12.3µs ± 0%    12.3µs ± 1%     ~     (p=0.214 n=5+5)
RangeQuery/expr=a_one,steps=10-12                                                                            12.8µs ± 2%    13.0µs ± 3%     ~     (p=0.310 n=5+5)
RangeQuery/expr=a_one,steps=100-12                                                                           19.3µs ± 1%    20.0µs ± 8%     ~     (p=0.841 n=5+5)
RangeQuery/expr=a_one,steps=1000-12                                                                          65.6µs ± 1%    68.9µs ± 2%   +4.92%  (p=0.008 n=5+5)
RangeQuery/expr=a_ten,steps=1-12                                                                             49.7µs ± 0%    49.7µs ± 0%     ~     (p=1.000 n=5+5)
RangeQuery/expr=a_ten,steps=10-12                                                                            53.8µs ± 0%    54.0µs ± 1%     ~     (p=0.151 n=5+5)
RangeQuery/expr=a_ten,steps=100-12                                                                            119µs ± 0%     119µs ± 0%     ~     (p=0.421 n=5+5)
RangeQuery/expr=a_ten,steps=1000-12                                                                           575µs ± 0%     602µs ± 2%   +4.59%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=1-12                                                                          439µs ± 0%     438µs ± 0%     ~     (p=0.286 n=4+5)
RangeQuery/expr=a_hundred,steps=10-12                                                                         506µs ±17%     481µs ± 0%     ~     (p=0.222 n=5+5)
RangeQuery/expr=a_hundred,steps=100-12                                                                       1.13ms ± 2%    1.13ms ± 0%     ~     (p=0.841 n=5+5)
RangeQuery/expr=a_hundred,steps=1000-12                                                                      5.68ms ± 2%    5.87ms ± 1%   +3.20%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=1-12                                                                   16.9µs ± 0%    16.8µs ± 0%   -0.97%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=10-12                                                                  18.0µs ± 0%    17.8µs ± 0%   -0.70%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=100-12                                                                 31.9µs ± 0%    31.7µs ± 0%     ~     (p=0.057 n=4+4)
RangeQuery/expr=rate(a_one[1m]),steps=1000-12                                                                 146µs ± 0%     149µs ± 2%   +1.99%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=1-12                                                                   54.1µs ± 0%    53.6µs ± 0%   -0.88%  (p=0.016 n=5+4)
RangeQuery/expr=rate(a_ten[1m]),steps=10-12                                                                  65.4µs ± 2%    64.6µs ± 1%   -1.25%  (p=0.032 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=100-12                                                                  204µs ± 2%     203µs ± 0%     ~     (p=0.690 n=5+5)
RangeQuery/expr=rate(a_ten[1m]),steps=1000-12                                                                1.34ms ± 0%    1.37ms ± 1%   +1.77%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-12                                                                433µs ± 1%     430µs ± 0%     ~     (p=0.095 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=10-12                                                               543µs ± 1%     542µs ± 0%     ~     (p=0.841 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-12                                                             1.94ms ± 2%    1.92ms ± 0%     ~     (p=0.548 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-12                                                            13.4ms ± 1%    13.6ms ± 0%   +1.15%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_one[1m]),steps=10000-12                                                               1.47ms ± 0%    1.48ms ± 0%   +0.41%  (p=0.016 n=5+4)
RangeQuery/expr=rate(a_ten[1m]),steps=10000-12                                                               14.6ms ± 0%    14.6ms ± 1%     ~     (p=0.056 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-12                                                            147ms ± 2%     146ms ± 1%     ~     (p=1.000 n=5+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-12                                                  764µs ± 2%     763µs ± 1%     ~     (p=0.841 n=5+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-12                                                1.24ms ± 0%    1.25ms ± 0%   +0.64%  (p=0.016 n=4+5)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-12                                               6.11ms ± 0%    6.12ms ± 0%     ~     (p=0.111 n=5+4)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-12                                              55.3ms ± 0%    55.4ms ± 1%     ~     (p=0.548 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-12                                                 6.03ms ± 0%    6.12ms ± 0%   +1.42%  (p=0.016 n=4+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-12                                                10.9ms ± 0%    11.0ms ± 0%   +0.60%  (p=0.008 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-12                                               60.1ms ± 0%    60.6ms ± 2%     ~     (p=0.056 n=5+5)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-12                                               542ms ± 2%     554ms ± 7%     ~     (p=0.095 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-12                                             60.7ms ± 2%    60.5ms ± 0%     ~     (p=1.000 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-12                                             110ms ± 0%     111ms ± 3%     ~     (p=0.151 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-12                                            592ms ± 2%     595ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-12                                           5.40s ± 0%     5.42s ± 0%   +0.37%  (p=0.016 n=4+5)
RangeQuery/expr=changes(a_one[1d]),steps=1-12                                                                 604µs ± 0%     608µs ± 0%   +0.73%  (p=0.008 n=5+5)
RangeQuery/expr=changes(a_one[1d]),steps=10-12                                                                766µs ± 0%     768µs ± 0%     ~     (p=0.063 n=4+5)
RangeQuery/expr=changes(a_one[1d]),steps=100-12                                                              2.42ms ± 2%    2.36ms ± 0%   -2.44%  (p=0.008 n=5+5)
RangeQuery/expr=changes(a_one[1d]),steps=1000-12                                                             19.0ms ± 1%    18.5ms ± 0%   -2.16%  (p=0.008 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=1-12                                                                5.29ms ± 0%    5.29ms ± 0%     ~     (p=0.841 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=10-12                                                               6.93ms ± 0%    6.90ms ± 0%     ~     (p=0.111 n=4+5)
RangeQuery/expr=changes(a_ten[1d]),steps=100-12                                                              23.8ms ± 2%    23.2ms ± 0%   -2.73%  (p=0.008 n=5+5)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-12                                                              187ms ± 1%     183ms ± 1%     ~     (p=0.095 n=5+5)
…