@roidelapluie 2021-03-13
I am wondering if we could, instead of pre-computing, pass the result of the hashes to the next step. That would solve the count_result impact and we could even maybe evict the hashes not used in the current step for high churn scenarios. Did you consider such an approach?