@roidelapluie 2021-03-13
To clarify:
- Our users complain more often about RAM usage than CPU usage. If we use more ram, prometheus will likely crash or be killed.
- Because I expect most of the count_values to run on instant queries, I don't think we should compute hashes twice here if there is a reasonable way not to do it.
- Our benchmarks do not have example with hich churn, but with 50k+ series I'd expect some churn here in practice.