@pracucci 2021-03-13
> I am wondering if we could, instead of pre-computing, pass the result of the hashes to the next step

To my understanding of your idea, I think would be less efficient. It's something I've considered initially (my idea was to add a specific "cache" to `EvalNodeHelper`, but the problem is that each step is not guaranteed to work on the whole set of series so in the 1st step you may compute the grouping key for "some" series and then, in the 2nd step, finding the series for which you miss would be inefficient. If you have a better idea I'm open to reconsider it.