@pracucci 2021-03-13
> Our users complain more often about RAM usage than CPU usage. If we use more ram, prometheus will likely crash or be killed.

This optimization increases the number of allocations, while the allocated memory is not much impacted. Moreover, I think Prometheus operators care about memory (and I agree) but Prometheus users (who run the queries) just want faster queries.