@pracucci 2021-03-13
> Because I expect most of the count_values to run on instant queries, I don't think we should compute hashes twice here if there is a reasonable way not to do it.

Let me add a benchmark on `count_values()` and then we can take an informed decision.