@roidelapluie 2021-03-13
The overhead for each serie seems minimal indeed, and I think that my concerns with high churn would not really matter in practice. I will review the code on Monday.