@pracucci 2021-03-15
> Let me add a benchmark on `count_values()` and then we can take an informed decision.

I've added it:

```
name                                                            old time/op    new time/op    delta
RangeQuery/expr=count_values('value',_h_one),steps=1-12           68.7µs ± 2%    69.6µs ± 1%    ~     (p=0.095 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=10-12           120µs ± 1%     123µs ± 3%  +2.77%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=100-12          677µs ± 2%     687µs ± 0%    ~     (p=0.151 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=1000-12        6.15ms ± 1%    6.26ms ± 1%  +1.72%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=1-12            567µs ± 2%     567µs ± 3%    ~     (p=1.000 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=10-12          1.04ms ± 1%    1.05ms ± 0%    ~     (p=0.063 n=5+4)
RangeQuery/expr=count_values('value',_h_ten),steps=100-12         6.15ms ± 1%    6.22ms ± 1%  +1.09%  (p=0.016 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=1000-12        59.5ms ±10%    56.7ms ± 1%    ~     (p=0.690 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1-12       6.40ms ± 8%    6.40ms ± 7%    ~     (p=0.690 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=10-12      11.2ms ± 2%    11.5ms ± 2%    ~     (p=0.056 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-12     61.5ms ± 1%    62.9ms ± 1%  +2.27%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1000-12     551ms ± 1%     564ms ± 0%  +2.35%  (p=0.016 n=5+4)

name                                                            old alloc/op   new alloc/op   delta
RangeQuery/expr=count_values('value',_h_one),steps=1-12           25.6kB ± 0%    26.0kB ± 0%  +1.56%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=10-12          56.6kB ± 0%    57.2kB ± 0%  +1.17%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=100-12          385kB ± 0%     388kB ± 0%  +0.87%  (p=0.016 n=4+5)
RangeQuery/expr=count_values('value',_h_one),steps=1000-12        3.77MB ± 0%    3.79MB ± 0%  +0.77%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=1-12            225kB ± 0%     228kB ± 0%  +1.24%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=10-12           584kB ± 0%     587kB ± 0%  +0.52%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=100-12         4.20MB ± 0%    4.20MB ± 0%  +0.13%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=1000-12        40.6MB ± 0%    40.6MB ± 0%    ~     (p=0.730 n=5+4)
RangeQuery/expr=count_values('value',_h_hundred),steps=1-12       2.18MB ± 0%    2.21MB ± 0%  +1.27%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=10-12      5.73MB ± 0%    5.75MB ± 0%  +0.51%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-12     41.4MB ± 0%    41.4MB ± 0%  +0.07%  (p=0.016 n=4+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1000-12     410MB ± 0%     410MB ± 0%    ~     (p=0.057 n=4+4)

name                                                            old allocs/op  new allocs/op  delta
RangeQuery/expr=count_values('value',_h_one),steps=1-12              436 ± 0%       455 ± 0%  +4.36%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=10-12           1.01k ± 0%     1.05k ± 0%  +3.66%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=100-12          6.83k ± 0%     7.04k ± 0%  +3.18%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_one),steps=1000-12         65.0k ± 0%     67.0k ± 0%  +3.03%  (p=0.029 n=4+4)
RangeQuery/expr=count_values('value',_h_ten),steps=1-12            3.01k ± 0%     3.13k ± 0%  +3.92%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=10-12           8.04k ± 0%     8.18k ± 0%  +1.69%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=100-12          58.8k ± 0%     59.1k ± 0%  +0.54%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_ten),steps=1000-12          568k ± 0%      569k ± 0%  +0.30%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1-12        28.8k ± 0%     29.9k ± 0%  +3.85%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=10-12       78.3k ± 0%     79.5k ± 0%  +1.44%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-12       579k ± 0%      580k ± 0%  +0.23%  (p=0.016 n=4+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1000-12     5.59M ± 0%     5.59M ± 0%  -0.02%  (p=0.029 n=4+4)
```

I'm not sure it's worth to add extra logic to skip the initial hashing (`prepSeries()`) for the case of `count_values()` given the small impact on it but, again, I'm open to your feedback. That's just my thought.