@pracucci 2021-03-17
I've deployed this change to one of our production clusters and I'm seeing a reduction between -10% and -15% on query ranges latency for queries using `sum by()`. Not impressive but, from this preliminary results, may still be worthwhile.